Refactoring tests
